+++ Scripts/Testers/__init__.py
@@ -1,1 +1,1 @@
-from Scripts.Testers.Scraper.Sports.Soccer.test_MatchReport import TestMatchReport
+