+++ Scripts/Utility/__init__.py
@@ -1,4 +1,0 @@
-from .path import *
-from .json import *
-from .logger import *
-from .Exceptions import *