+++ Scripts/__init__.py
@@ -1,4 +1,0 @@
-from .Handlers import *
-from .Scraper import *
-from .Testers import *
-from .Utility import *